add progress for appends
