+++ commands/share.js
@@ -2,12 +2,16 @@ var chalk = require('chalk')
 var prettyBytes = require('pretty-bytes')
 var Dat = require('../lib/dat')
 var logger = require('../lib/logger')
+var ui = require('../lib/ui')
 
 module.exports = function (args) {
   var dat = Dat(args)
   var log = logger(args)
 
   var addText = 'Adding '
+  var finalized = false
+  var updated = false
+  var initFileCount = 0
 
   log.status('Starting Dat...\n', 0)
   log.status('Connecting...', 1)
@@ -28,13 +32,20 @@ module.exports = function (args) {
   dat.on('file-added', printStats)
   dat.on('file-exists', printStats)
 
+  dat.on('append-ready', function () {
+    printStats()
+  })
+
   dat.on('archive-finalized', function () {
     addText = 'Added '
+    finalized = true
+    initFileCount = dat.stats.filesTotal
     printStats()
   })
 
   dat.on('archive-updated', function () {
     addText = 'Updated '
+    updated = true
     printStats()
   })
 
@@ -56,8 +67,11 @@ module.exports = function (args) {
 
   function printStats (data) {
     var stats = dat.stats
-    var msg = addText + chalk.bold(stats.filesTotal) + ' files'
-    msg += chalk.dim(' (' + prettyBytes(stats.bytesTotal) + ')')
+    var files = stats.filesTotal
+    if (updated) files = files - initFileCount
+    var msg = ui.progress(stats.bytesTotal / dat.appendStats.bytes)
+    msg += ' ' + addText + chalk.bold(files) + ' files'
+    msg += chalk.dim(' (' + prettyBytes(stats.bytesTotal) + '/' + prettyBytes(dat.appendStats.bytes) + ')')
     log.status(msg + '\n', 0)
   }
 }
