+++ lib/append.js
@@ -1,15 +1,23 @@
+var fs = require('fs')
+var path = require('path')
+var countFiles = require('count-files')
 var walker = require('folder-walker')
 var each = require('stream-each')
 
 module.exports.initialAppend = function (dat, cb) {
-  if (dat.resume) {
-    dat.db.get('!dat!finalized', function (err, val) {
-      if (err || val !== 'true') walkFolder(true)
-      else walkFolder(true) // TODO: check mtimes
-    })
-  } else {
-    walkFolder()
-  }
+  countFiles(dat.dir, {filter: ignore}, function (err, stats) {
+    dat.appendStats = stats
+    dat.emit('append-ready')
+
+    if (dat.resume) {
+      dat.db.get('!dat!finalized', function (err, val) {
+        if (err || val !== 'true') walkFolder(true)
+        else walkFolder(true) // TODO: check mtimes
+      })
+    } else {
+      walkFolder()
+    }
+  })
 
   function walkFolder (resume) {
     var fileStream = walker(dat.dir, {filter: ignore})
