+++ lib/dat.js
@@ -108,7 +108,6 @@ Dat.prototype.share = function (cb) {
 
       var watch = yoloWatch(self.dir)
       watch.on('changed', function (name, data) {
-        // TODO: make yolowatch data consistent w/ folder-walker
         if (name === self.dir) return
         append.liveAppend(self, data)
         self.emit('archive-updated')
