+++ package.json
@@ -24,6 +24,7 @@
   "license": "BSD-3-Clause",
   "dependencies": {
     "chalk": "^1.1.1",
+    "count-files": "^1.0.0",
     "datland-swarm-defaults": "^1.0.0",
     "debug": "^2.2.0",
     "discovery-swarm": "^4.0.0",
